# frontend/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document Archive API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/ingest/telegram": {
      "post": {
        "summary": "Ingest Telegram",
        "operationId": "ingest_telegram_api_ingest_telegram_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_telegram_api_ingest_telegram_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestAcceptedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingest/telegram/batch": {
      "post": {
        "summary": "Ingest Telegram Batch",
        "operationId": "ingest_telegram_batch_api_ingest_telegram_batch_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_telegram_batch_api_ingest_telegram_batch_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestBatchAcceptedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingest/actions/{job_id}/retry": {
      "post": {
        "summary": "Ingest Action Retry",
        "operationId": "ingest_action_retry_api_ingest_actions__job_id__retry_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          },
          {
            "name": "X-OpenClaw-Action-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Openclaw-Action-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/IngestActionRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingest/actions/{job_id}/reprocess": {
      "post": {
        "summary": "Ingest Action Reprocess",
        "operationId": "ingest_action_reprocess_api_ingest_actions__job_id__reprocess_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Token"
            }
          },
          {
            "name": "X-OpenClaw-Action-Token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Openclaw-Action-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/IngestActionRequest"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Req"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestActionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingest/manual": {
      "post": {
        "summary": "Ingest Manual",
        "operationId": "ingest_manual_api_ingest_manual_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_manual_api_ingest_manual_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestAcceptedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ingest/manual/batch": {
      "post": {
        "summary": "Ingest Manual Batch",
        "operationId": "ingest_manual_batch_api_ingest_manual_batch_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_ingest_manual_batch_api_ingest_manual_batch_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestBatchAcceptedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_ingest_manual_api_ingest_manual_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "source": {
            "$ref": "#/components/schemas/SourceType",
            "default": "manual"
          },
          "source_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_ingest_manual_api_ingest_manual_post"
      },
      "Body_ingest_manual_batch_api_ingest_manual_batch_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          },
          "source": {
            "$ref": "#/components/schemas/SourceType",
            "default": "manual"
          },
          "source_ref_prefix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref Prefix"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "required": [
          "files"
        ],
        "title": "Body_ingest_manual_batch_api_ingest_manual_batch_post"
      },
      "Body_ingest_telegram_api_ingest_telegram_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "source": {
            "$ref": "#/components/schemas/SourceType",
            "default": "telegram"
          },
          "source_ref": {
            "type": "string",
            "title": "Source Ref"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "chat_id": {
            "type": "string",
            "title": "Chat Id"
          },
          "sent_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          }
        },
        "type": "object",
        "required": [
          "file",
          "source_ref",
          "message_id",
          "chat_id"
        ],
        "title": "Body_ingest_telegram_api_ingest_telegram_post"
      },
      "Body_ingest_telegram_batch_api_ingest_telegram_batch_post": {
        "properties": {
          "files": {
            "items": {
              "type": "string",
              "contentMediaType": "application/octet-stream"
            },
            "type": "array",
            "title": "Files"
          },
          "source": {
            "$ref": "#/components/schemas/SourceType",
            "default": "telegram"
          },
          "source_ref_prefix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref Prefix"
          },
          "message_id": {
            "type": "string",
            "title": "Message Id"
          },
          "chat_id": {
            "type": "string",
            "title": "Chat Id"
          },
          "sent_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At"
          },
          "caption": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption"
          }
        },
        "type": "object",
        "required": [
          "files",
          "message_id",
          "chat_id"
        ],
        "title": "Body_ingest_telegram_batch_api_ingest_telegram_batch_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "IngestAcceptedResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "state": {
            "$ref": "#/components/schemas/IngestState"
          },
          "source": {
            "$ref": "#/components/schemas/SourceType"
          },
          "source_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Ref"
          },
          "queued_at": {
            "type": "string",
            "format": "date-time",
            "title": "Queued At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "state",
          "source",
          "source_ref",
          "queued_at"
        ],
        "title": "IngestAcceptedResponse"
      },
      "IngestActionRequest": {
        "properties": {
          "reset_attempts": {
            "type": "boolean",
            "title": "Reset Attempts",
            "default": false
          },
          "clear_error": {
            "type": "boolean",
            "title": "Clear Error",
            "default": true
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          },
          "caption_override": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caption Override"
          }
        },
        "type": "object",
        "title": "IngestActionRequest"
      },
      "IngestActionResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "format": "uuid",
            "title": "Job Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "previous_state": {
            "$ref": "#/components/schemas/IngestState"
          },
          "state": {
            "$ref": "#/components/schemas/IngestState"
          },
          "enqueued": {
            "type": "boolean",
            "title": "Enqueued"
          },
          "queued_at": {
            "type": "string",
            "format": "date-time",
            "title": "Queued At"
          },
          "attempt_count": {
            "type": "integer",
            "title": "Attempt Count"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "action",
          "previous_state",
          "state",
          "enqueued",
          "queued_at",
          "attempt_count"
        ],
        "title": "IngestActionResponse"
      },
      "IngestBatchAcceptedResponse": {
        "properties": {
          "total_files": {
            "type": "integer",
            "title": "Total Files"
          },
          "accepted_count": {
            "type": "integer",
            "title": "Accepted Count"
          },
          "rejected_count": {
            "type": "integer",
            "title": "Rejected Count"
          },
          "accepted": {
            "items": {
              "$ref": "#/components/schemas/IngestAcceptedResponse"
            },
            "type": "array",
            "title": "Accepted"
          },
          "rejected": {
            "items": {
              "$ref": "#/components/schemas/IngestBatchRejectedItem"
            },
            "type": "array",
            "title": "Rejected"
          }
        },
        "type": "object",
        "required": [
          "total_files",
          "accepted_count",
          "rejected_count"
        ],
        "title": "IngestBatchAcceptedResponse"
      },
...
      "IngestState": {
        "type": "string",
        "enum": [
          "RECEIVED",
          "STORED",
          "EXTRACTED",
          "CLASSIFIED",
          "INDEXED",
          "PUBLISHED",
          "FAILED",
          "NEEDS_REVIEW"
        ],
        "title": "IngestState"
      },
...
      "SourceType": {
        "type": "string",
        "enum": [
          "telegram",
          "wiki",
          "manual",
          "api"
        ],
        "title": "SourceType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}